packages/easyobs-agent/pyproject.toml
[project]
name = "easyobs-agent"
version = "0.1.0"
description = "EasyObs SDK: lightweight LLM/agent observability instrumentation via OpenTelemetry"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "EasyObs Team" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries",
  "Topic :: System :: Monitoring",
]
keywords = ["observability", "llm", "opentelemetry", "tracing", "langchain", "agent"]
dependencies = [
  "opentelemetry-api>=1.28.0",
  "opentelemetry-sdk>=1.28.0",
  "opentelemetry-exporter-otlp-proto-http>=1.28.0",
  "langchain-core>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/syyoo/easyobs-agent"
Documentation = "https://github.com/syyoo/easyobs-agent#readme"
Repository = "https://github.com/syyoo/easyobs-agent"
Issues = "https://github.com/syyoo/easyobs-agent/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/easyobs_agent"]

[tool.ruff]
line-length = 100
target-version = "py39"
